refactor: form reuse promotion
