+++ maid-care-service-frontend/src/components/pages/promotion/constants/form-mode.js
@@ -1,0 +1,4 @@
+export const FORM_MODE = {
+  CREATE: 'CREATE',
+  EDIT: 'EDIT',
+};
